Vulnerabilities/2019/CVE-2019-20933/Hydragyrum/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# CVE-2019-20933/influx-client.py: 6
PyJWT == 2.0.1

# CVE-2019-20933/influx-client.py: 5
requests == 2.25.1
